refactor: reorganize project structure for session module
