+++ app/crud/user.py
@@ -1,4 +1,3 @@
-from datetime import datetime
 from sqlalchemy.orm import Session
 from app import models, schemas, auth
 
@@ -31,28 +30,3 @@ def get_user(db: Session, user_id: int):
 
 def get_users(db: Session, skip: int = 0, limit: int = 100):
     return db.query(models.User).offset(skip).limit(limit).all()
-
-
-def create_session(db: Session, session_data: schemas.SessionCreate, owner_id: int):
-    invited_users = db.query(models.User).filter(models.User.id.in_(session_data.invited_users_ids)).all()
-    new_session = models.Session(
-        name=session_data.name,
-        location_radius=session_data.location_radius,
-        owner_id=owner_id,
-        invited_users=invited_users,
-        created_at=datetime.utcnow()
-
-    )
-    db.add(new_session)
-    db.commit()
-    db.refresh(new_session)
-    return new_session
-
-def get_sessions_for_user(db: Session, user_id: int):
-    user = db.query(models.User).filter(models.User.id == user_id).first()
-    if not user:
-        return []
-
-    sessions = user.owned_sessions + user.invited_sessions
-    return sorted(sessions, key=lambda s: s.created_at or datetime.min,  reverse=True)
-
